requirements.txt
python-telegram-bot==20.8
pillow==11.2.1